fix(SegmentList): only show tooltips on Part Title overflow
